requirements_enhanced.txt
# Core dependencies
customtkinter>=5.2.0
psutil>=5.9.0
pillow>=9.0.0
numpy>=1.21.0

# Media processing
opencv-python>=4.5.0
pygame>=2.1.0
sounddevice>=0.4.0
scipy>=1.7.0

# System utilities
beautifulsoup4>=4.10.0
requests>=2.25.0
py-cpuinfo>=8.0.0
keyboard>=0.13.0

# Windows specific (install only on Windows)
wmi>=1.5.1; sys_platform == "win32"
pywifi>=1.1.12; sys_platform == "win32"
pywin32>=227; sys_platform == "win32"

# Enhanced features - PDF/Excel export
reportlab>=3.6.0
pandas>=1.3.0
openpyxl>=3.0.9

# Optional - for better performance
numba>=0.56.0
matplotlib>=3.5.0